darkmode toggle mostly works.
Cookie needs to be fixed so on refresh/load the theme stays the same instead of resetting to default.

homepage containers still have issue exapnding the background image to the appropriate size.

the toggle doesnt change even though the value of checkbox does.

the toggle and checkbox both are not formatted perfectly